+++ sdk/storage/azure-storage-blob/azure/storage/blob/aio/blob_service_client_async.py
@@ -35,7 +35,7 @@ if TYPE_CHECKING:
     from .lease_async import LeaseClient
     from ..models import (
         BlobProperties,
-        Logging,
+        BlobAnalyticsLogging,
         Metrics,
         RetentionPolicy,
         StaticWebsite,
@@ -235,7 +235,7 @@ class BlobServiceClient(AsyncStorageAccountHostsMixin, BlobServiceClientBase):
 
     @distributed_trace_async
     async def set_service_properties(
-            self, logging=None,  # type: Optional[Logging]
+            self, logging=None,  # type: Optional[BlobAnalyticsLogging]
             hour_metrics=None,  # type: Optional[Metrics]
             minute_metrics=None,  # type: Optional[Metrics]
             cors=None,  # type: Optional[List[CorsRule]]
@@ -248,12 +248,12 @@ class BlobServiceClient(AsyncStorageAccountHostsMixin, BlobServiceClientBase):
         """Sets the properties of a storage account's Blob service, including
         Azure Storage Analytics.
 
-        If an element (e.g. Logging) is left as None, the
+        If an element (e.g. BlobAnalyticsLogging) is left as None, the
         existing settings on the service for that functionality are preserved.
 
         :param logging:
             Groups the Azure Analytics Logging settings.
-        :type logging: ~azure.storage.blob.Logging
+        :type logging: ~azure.storage.blob.BlobAnalyticsLogging
         :param hour_metrics:
             The hour metrics settings provide a summary of request
             statistics grouped by API in hourly aggregates for blobs.