+++ sdk/storage/azure-storage-blob/azure/storage/blob/blob_service_client.py
@@ -38,7 +38,7 @@ if TYPE_CHECKING:
     from .lease import LeaseClient
     from .models import (
         BlobProperties,
-        Logging,
+        BlobAnalyticsLogging,
         Metrics,
         RetentionPolicy,
         StaticWebsite,
@@ -354,7 +354,7 @@ class BlobServiceClient(StorageAccountHostsMixin):
 
     @distributed_trace
     def set_service_properties(
-            self, logging=None,  # type: Optional[Logging]
+            self, logging=None,  # type: Optional[BlobAnalyticsLogging]
             hour_metrics=None,  # type: Optional[Metrics]
             minute_metrics=None,  # type: Optional[Metrics]
             cors=None,  # type: Optional[List[CorsRule]]
@@ -367,12 +367,12 @@ class BlobServiceClient(StorageAccountHostsMixin):
         """Sets the properties of a storage account's Blob service, including
         Azure Storage Analytics.
 
-        If an element (e.g. Logging) is left as None, the
+        If an element (e.g. BlobAnalyticsLogging) is left as None, the
         existing settings on the service for that functionality are preserved.
 
         :param logging:
             Groups the Azure Analytics Logging settings.
-        :type logging: ~azure.storage.blob.Logging
+        :type logging: ~azure.storage.blob.BlobAnalyticsLogging
         :param hour_metrics:
             The hour metrics settings provide a summary of request
             statistics grouped by API in hourly aggregates for blobs.