+++ sdk/storage/azure-storage-queue/azure/storage/queue/aio/queue_service_client_async.py
@@ -36,7 +36,7 @@ if TYPE_CHECKING:
     from azure.storage.queue.aio.models import (
         QueueProperties
     )
-    from azure.storage.queue.models import Logging, Metrics, CorsRule
+    from azure.storage.queue.models import QueueAnalyticsLogging, Metrics, CorsRule
 
 
 class QueueServiceClient(AsyncStorageAccountHostsMixin, QueueServiceClientBase):
@@ -167,7 +167,7 @@ class QueueServiceClient(AsyncStorageAccountHostsMixin, QueueServiceClientBase):
 
     @distributed_trace_async
     async def set_service_properties( # type: ignore
-            self, logging=None,  # type: Optional[Logging]
+            self, logging=None,  # type: Optional[QueueAnalyticsLogging]
             hour_metrics=None,  # type: Optional[Metrics]
             minute_metrics=None,  # type: Optional[Metrics]
             cors=None,  # type: Optional[List[CorsRule]]
@@ -177,12 +177,12 @@ class QueueServiceClient(AsyncStorageAccountHostsMixin, QueueServiceClientBase):
         """Sets the properties of a storage account's Queue service, including
         Azure Storage Analytics.
 
-        If an element (e.g. Logging) is left as None, the
+        If an element (e.g. QueueAnalyticsLogging) is left as None, the
         existing settings on the service for that functionality are preserved.
 
         :param logging:
             Groups the Azure Analytics Logging settings.
-        :type logging: ~azure.storage.queue.Logging
+        :type logging: ~azure.storage.queue.QueueAnalyticsLogging
         :param hour_metrics:
             The hour metrics settings provide a summary of request
             statistics grouped by API in hourly aggregates for queues.