+++ sdk/storage/azure-storage-blob/HISTORY.md
@@ -6,6 +6,7 @@
 
 - `set_container_access_policy` has required parameter `signed_identifiers`.
 - NoRetry policy has been removed. Use keyword argument `retry_total=0` for no retries. 
+- `Logging` has been renamed to `BlobAnalyticsLogging`.
 
 **New features**