+++ sdk/storage/azure-storage-queue/HISTORY.md
@@ -15,6 +15,7 @@ To use a queue_url, the method `from_queue_url` must be used.
   - `timeout` etc.
 - `QueueMessage` has had its parameters renamed from `insertion_time`, `time_next_visible`, `expiration_time`
 to `inserted_on`, `next_visible_on`, `expires_on`, respectively. 
+- `Logging` has been renamed to `QueueAnalyticsLogging`.
 
  **New features**